# Conflicts:
#	cli/src/main/java/com/devonfw/tools/ide/tool/LocalToolCommandlet.java

+++ cli/src/main/java/com/devonfw/tools/ide/tool/GlobalToolCommandlet.java
@@ -24,7 +24,7 @@ public abstract class GlobalToolCommandlet extends ToolCommandlet {
    * @param context the {@link IdeContext}.
    * @param tool the {@link #getName() tool name}.
    * @param tags the {@link #getTags() tags} classifying the tool. Should be created via {@link Set#of(Object) Set.of}
-   *        method.
+   * method.
    */
   public GlobalToolCommandlet(IdeContext context, String tool, Set<Tag> tags) {
 
@@ -56,17 +56,20 @@ public abstract class GlobalToolCommandlet extends ToolCommandlet {
     // download and install the global tool
     FileAccess fileAccess = this.context.getFileAccess();
     Path target = toolRepository.download(this.tool, edition, resolvedVersion);
-    Path tmpDir = fileAccess.createTempDir(getName());
-    Path downloadBinaryPath = tmpDir.resolve(target.getFileName());
-    extract(target, downloadBinaryPath);
-    if (isExtract()) {
+    Path executable = target;
+    Path tmpDir = null;
+    boolean extract = isExtract();
+    if (extract) {
+      tmpDir = fileAccess.createTempDir(getName());
+      Path downloadBinaryPath = tmpDir.resolve(target.getFileName());
+      fileAccess.extract(target, downloadBinaryPath);
       downloadBinaryPath = fileAccess.findFirst(downloadBinaryPath, Files::isExecutable, false);
     }
-    ProcessContext pc = this.context.newProcess().errorHandling(ProcessErrorHandling.WARNING)
-        .executable(downloadBinaryPath);
+    ProcessContext pc = this.context.newProcess().errorHandling(ProcessErrorHandling.WARNING).executable(executable);
     int exitCode = pc.run();
-    fileAccess.delete(tmpDir);
-    fileAccess.delete(target);
+    if (tmpDir != null) {
+      fileAccess.delete(tmpDir);
+    }
     if (exitCode == 0) {
       this.context.success("Successfully installed {} in version {}", this.tool, resolvedVersion);
     } else {
